+++ BossMod/Modules/Heavensward/Dungeon/D06TheAetherochemicalResearchFacility/D061RegulaVanHydrus.cs
@@ -33,7 +33,7 @@ class AetherochemicalGrenado(BossModule module) : Components.LocationTargetedAOE
 class AetherochemicalLaser(BossModule module) : Components.BaitAwayTethers(module, new AOEShapeRect(50, 2), (uint)TetherID.BaitAway, ActionID.MakeSpell(AID.AetherochemicalLaser));
 class AetherochemicalLaserAOE(BossModule module) : Components.SelfTargetedAOEs(module, ActionID.MakeSpell(AID.AetherochemicalLaser), new AOEShapeRect(50, 2));
 class JudgmentAOE(BossModule module) : Components.SelfTargetedAOEs(module, ActionID.MakeSpell(AID.JudgmentAOE), new AOEShapeCircle(8));
-class MagiteckTurrents(BossModule module) : Components.AddsMulti(module, [(uint)OID.MagitekTurretI, (uint)OID.MagitekTurretII]);
+class MagiteckTurrents(BossModule module) : Components.AddsMulti(module, [OID.MagitekTurretI, OID.MagitekTurretII]);
 class MagitekSlug(BossModule module) : Components.SelfTargetedAOEs(module, ActionID.MakeSpell(AID.MagitekSlug), new AOEShapeRect(61.65f, 2));
 class MagitekSpread(BossModule module) : Components.SelfTargetedAOEs(module, ActionID.MakeSpell(AID.MagitekSpread), new AOEShapeCone(31.65f, 120.Degrees()));
 class SelfDetonate(BossModule module) : Components.RaidwideCast(module, ActionID.MakeSpell(AID.SelfDetonate));
